+++ intake_thredds/cat.py
@@ -27,11 +27,6 @@ class ThreddsCatalog(Catalog):
     def __init__(self, url: str, driver: str = 'opendap', **kwargs):
         self.url = url
         self.driver = driver
-        if "decode_times" in kwargs:
-            self.xarray_kwargs = {"xarray_kwargs": {"decode_times": kwargs["decode_times"]}}
-            del kwargs["decode_times"]
-        else:
-            self.xarray_kwargs = {}
         super().__init__(**kwargs)
 
     def _load(self):